gitlab-ci-memcached/requirements.txt
python-memcached==1.58
pytest==4.3.0
